api/requirements.txt
requests
json
datetime 
wordcloud 
transformers 
matplotlib.pyplot 
numpy 